Add security authentication, and employee form for add

+++ src/main/java/com/springapp/carrepairshop/carrepairshop/entity/Role.java
@@ -9,7 +9,7 @@ public class Role
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Long id;
+    private int id;
 
     @Column(name = "name")
     private String name;
@@ -23,12 +23,12 @@ public class Role
         this.name = name;
     }
 
-    public Long getId()
+    public int getId()
     {
         return id;
     }
 
-    public void setId(Long id)
+    public void setId(int id)
     {
         this.id = id;
     }
